Move Endpoint, Controller and Router to separate files
And add a TestDispatchForm.ConnCase module where we handle the imports,
set the endpoint, and build a conn.
