+++ test/test_dispatch_form_test.exs
@@ -1,63 +1,5 @@
-defmodule TestDispatchFormTest.Controller do
-  import Plug.Conn, only: [put_resp_content_type: 2, resp: 3]
-
-  def init(opts), do: opts
-
-  def call(conn, :new) do
-    body = File.read!("test/forms/_entity_and_form_controls.html")
-    conn |> put_resp_content_type("text/html") |> resp(200, body)
-  end
-
-  def call(%{params: %{"user" => %{"name" => _, "email" => _, "roles" => _}}} = conn, :create) do
-    conn |> put_resp_content_type("text/html") |> resp(200, "user created")
-  end
-end
-
-defmodule TestDispatchFormTest.Router do
-  use Phoenix.Router
-  alias TestDispatchFormTest.Controller
-
-  pipeline :browser do
-    plug(:put_bypass, :browser)
-  end
-
-  scope "/users" do
-    pipe_through(:browser)
-    get("/new", Controller, :new)
-    post("/create", Controller, :create)
-  end
-
-  def put_bypass(conn, pipeline) do
-    bypassed = (conn.assigns[:bypassed] || []) ++ [pipeline]
-    Plug.Conn.assign(conn, :bypassed, bypassed)
-  end
-end
-
 defmodule TestDispatchFormTest do
-  use ExUnit.Case, async: true
-
-  import Phoenix.ConnTest
-  import TestDispatchForm
-
-  alias TestDispatchFormTest.Router
-
-  Application.put_env(:test_dispatch_form, TestDispatchFormTest.Endpoint, [])
-
-  defmodule Endpoint do
-    use Phoenix.Endpoint, otp_app: :test_dispatch_form
-    def init(opts), do: opts
-
-    def call(conn, opts),
-      do: super(conn, opts).private[:endpoint] |> put_in(opts) |> Router.call(Router.init([]))
-  end
-
-  @endpoint Endpoint
-
-  setup do
-    Endpoint.start_link()
-    Logger.disable(self())
-    {:ok, conn: build_conn()}
-  end
+  use TestDispatchForm.ConnCase
 
   describe "form with entity and empty form controls" do
     test "dispatches form with attributes", %{conn: conn} do
